+++ src/Action/ActionInstance.php
@@ -27,7 +27,7 @@ class ActionInstance extends Model
      * @var array
      */
     protected $fillable = [
-        'name', 'description', 'event', 'action', 'module_instance_id', 'user_id'
+        'name', 'description', 'event', 'action', 'module_instance_id', 'user_id', 'should_queue'
     ];
 
     /**